refactor(condo): DOMA-9362 moved variables to the folder with corresponding integration
