+++ apps/condo/domains/organization/integrations/sbbol/sync/requestTransactions.js
@@ -8,11 +8,15 @@ const { getSchemaCtx } = require('@open-condo/keystone/schema')
 const { BANK_INTEGRATION_IDS, BANK_SYNC_TASK_STATUS } = require('@condo/domains/banking/constants')
 const { BankAccount, BankTransaction, BankContractorAccount, predictTransactionClassification, BankSyncTask, BankIntegrationAccountContext } = require('@condo/domains/banking/utils/serverSchema')
 const { RUSSIA_COUNTRY } = require('@condo/domains/common/constants/countries')
-const { ISO_CODES, ALTERNATIVE_CURRENCY_CODES_FROM_SBBOL } = require('@condo/domains/common/constants/currencies')
-const { dvSenderFields, INVALID_DATE_RECEIVED_MESSAGE } = require('@condo/domains/organization/integrations/sbbol/constants')
-const { SBBOL_IMPORT_NAME } = require('@condo/domains/organization/integrations/sbbol/constants')
-const { ERROR_PASSED_DATE_IN_THE_FUTURE } = require('@condo/domains/organization/integrations/sbbol/constants')
-const { SBBOL_ERRORS } = require('@condo/domains/organization/integrations/sbbol/constants')
+const { ISO_CODES } = require('@condo/domains/common/constants/currencies')
+const {
+    ALTERNATIVE_CURRENCY_CODES_FROM_SBBOL,
+    dvSenderFields,
+    INVALID_DATE_RECEIVED_MESSAGE,
+    SBBOL_IMPORT_NAME,
+    ERROR_PASSED_DATE_IN_THE_FUTURE,
+    SBBOL_ERRORS,
+} = require('@condo/domains/organization/integrations/sbbol/constants')
 const { initSbbolFintechApi, initSbbolClientWithToken } = require('@condo/domains/organization/integrations/sbbol/SbbolFintechApi')
 const { getAllAccessTokensByOrganization } = require('@condo/domains/organization/integrations/sbbol/utils/getAccessTokenForUser')
